KSB-1-3 fix data type

+++ src/main/kotlin/lesson1/lesson1_task4.kt
@@ -3,7 +3,7 @@ package org.example.lesson1
 fun main() {
 
     val length: Long = 40868600000
-    val age: Int = 27
+    val age: Byte = 27
     val partOfDay: Float = 0.075F
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
